Sync GMCM key bindings on startup.
Since field names are not even close to being unique, tried to disambiguate better (both for display and for storage) by looking at pages and headings. Field ID is nearly useless since mod authors don't specify it.

Seems to do a pretty decent job of dealing with the mess now. Still need to implement sync-on-change.