Add appearance settings to GMCM.
Includes a live preview for colors only. Most users shouldn't mess with the sizes, and reflecting them in a sprite-based preview is too hard right now. (And drawing the actual menu inside the GMCM isn't really feasible.)

Note: Contains a reference to the GMCMOptions mod, but support hasn't been added yet. Need to add this.