Add TextureSegment option to GMCM item editor.
In addition to adding the required controls, this adds another pile of hacks to show or hide the different editing rows depending on which image type is selected. The most difficult and fragile part is probably eliminating the extra "padding" whitespace for "hidden" rows, but this basically works.

The menu probably needs more testing but it seems to work close to 100% now. No crashes, UI updates, items are saved correctly, etc.