Limit activation sound to delayed actions.
For immediate activations, it's annoying as it often plays simultaneously with another sound, and just feels wrong with some mod shortcuts.

Also a bit of a performance tweak as this now won't keep trying to activate the item again on every frame until the delay is up. It will try once, then when delayed, stop trying until the delay expires.