More flexible activation settings.
Instead of an explicit "activate to consume/use, alt button to select", this sets up explicit primary and secondary activation methods, each with their own preference for what to do.

Code-wise, this is little more than adding a couple of new settings and changing the recently-added boolean to an enum.

Includes screenshot and documentation update.