Fix tinting and descriptions for "special" items.
Some flavored/preserved items, like Roe, have a special tinting method using an overlay; other items use just a color tint; and Smoked Fish are snowflakes that do a bit of both (and are implemented that way in the game itself).

There doesn't seem to be a very reliable way to get a _complete_ description from the `ParsedItemData`, so this instead adds a hack to try to reconstitute it from the result of `getDescription()`. It probably has other tradeoffs, as there's no way to know whether any line break is explicit or due to default wrapping, but we can make a few reasonable assumptions and get an OK result.