Support next/prev page while menu is open.
L/R shoulder buttons can be used to switch the page (i.e. backpack page, starting offset) displayed in the radial menu, while the menu is open. This means not having to dismiss the menu, press L/R shoulder buttons to change backpack page in vanilla, and then reopen the menu.

Offset is intentionally *local* to the radial menu so that cancelling out of the menu or activating a consumable item while on a non-"current" page does not change the current tool selection.