Highlight the current tool selection in menu.
Another QoL feature - when using the menu in reflex situations or as a form of pause-buffering, it's easy to lose track of which tool is actually selected when the game doesn't draw it overhead, and requires dismissing the menu and visually tracking the vanilla toolbar.

This adds separate highlighting (configurable color) for the currently-selected tool, making it visually distinct from both the highlighted/focused item and other non-focused/non-selected items.

Should work correctly with the shoulder-button page-flipping just implemented. Assumes that the current tool selection is always on the first page, which should always be the case unless another mod has broken it.