Refactor and add minor config tweaks for extensibility.
Menus, pages and items are now largely accessed and controlled through a hierarchy of interfaces, which can also be exposed through an API (though they aren't, yet).

Menus can also be configured to remember the previous page they were on instead of resetting. The default is still to reset, but memory may be useful if a downstream mod wants to add many pages.

This also happens to fix one of the earlier design flaws where items picked up while the menu is open wouldn't show up until closing/reopening the menu, or flipping back and forth between pages. Because items are essentially persistent (invalidated by events), not built on the fly, we can reassign them to the Painter on every frame since it is just a simple reference assignment.

A final config tweak allows custom items to specify their own behavior. We need this for mod-added items because there's no way to predict whether the mod's function "wants" a delay; since it's required anyway, adding it as a config option for GMCM-configured items is logical (e.g. NPC Compass or Data Layers which toggle a mode).