First try at menu registration API.
Includes registration and invalidation, both implicit by config and explicit by API. Integration is a little awkward due to the need to pass lists around (else we'd have to waste a lot of CPU and memory recreating things all the time) but seems to work - at least it does not break any of the existing functionality, and SMAPI/Pintail are OK with the API definition.

However, none of this has been tested; we'll need an example mod to try it out.