Check for null texture when drawing selection details.
Behavior seems to vary by installation; for some people a null texture causes this to fail and logs an error to SMAPI console, for others (possibly due a Harmony patch from some other mod) it is simply skipped with no unwanted side effect.

To be safe/consistent for everyone, don't try to draw a null texture in the selection details. This logic is only needed for selection since the main list has explicit monogram fallback logic. Selection details doesn't (on purpose - it looks bad) but should draw nothing when no image is available.